+++ src/search8.cc
@@ -949,7 +949,7 @@ void search8(BYTE * * q_start,
                       seq_id[c] = next_id;
                       int64_t seqno = seqnos[next_id];
                       char* address;
-                      int64_t length;
+                      unsigned int length;
 
                       db_getsequenceandlength(seqno, & address, & length);
